Split first migration
