+++ sdk/src/wallet/migration/mod.rs
@@ -2,13 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 
 mod migrate_0;
+mod migrate_1;
 
 use std::collections::HashMap;
 
 use anymap::Map;
 use async_trait::async_trait;
 use once_cell::sync::Lazy;
-use serde::{Deserialize, Serialize};
+use serde::{de::DeserializeOwned, Deserialize, Serialize};
 
 use crate::{
     client::storage::StorageAdapter,
@@ -25,24 +26,26 @@ static MIGRATIONS: Lazy<Map<dyn anymap::any::Any + Send + Sync>> = Lazy::new(|| 
     #[cfg(feature = "storage")]
     {
         use super::storage::Storage;
-        const STORAGE_MIGRATIONS: [(Option<usize>, &'static dyn DynMigration<Storage>); 1] = [
+        const STORAGE_MIGRATIONS: [(Option<usize>, &'static dyn DynMigration<Storage>); 2] = [
             // In order to add a new storage migration, add an entry at the bottom of this list
             // and change the list length above.
             // The entry should be in the form of a key-value pair, from previous migration to next.
             // i.e. (Some(migrate_<N>::Migrate::ID), &migrate_<N+1>::Migrate)
             (None, &migrate_0::Migrate),
+            (Some(migrate_0::Migrate::ID), &migrate_1::Migrate),
         ];
         migrations.insert(std::collections::HashMap::from(STORAGE_MIGRATIONS));
     }
     #[cfg(feature = "stronghold")]
     {
         use crate::client::stronghold::StrongholdAdapter;
-        const BACKUP_MIGRATIONS: [(Option<usize>, &'static dyn DynMigration<StrongholdAdapter>); 1] = [
+        const BACKUP_MIGRATIONS: [(Option<usize>, &'static dyn DynMigration<StrongholdAdapter>); 2] = [
             // In order to add a new backup migration, and add an entry at the bottom of this list
             // and change the list length above.
             // The entry should be in the form of a key-value pair, from previous migration to next.
             // i.e. (Some(migrate_<N>::Migrate::ID), &migrate_<N+1>::Migrate)
             (None, &migrate_0::Migrate),
+            (Some(migrate_0::Migrate::ID), &migrate_1::Migrate),
         ];
         migrations.insert(LatestBackupMigration(BACKUP_MIGRATIONS.last().unwrap().1.version()));
         migrations.insert(std::collections::HashMap::from(BACKUP_MIGRATIONS));
@@ -142,4 +145,37 @@ fn migrations<S: 'static + StorageAdapter>(
 #[cfg(feature = "stronghold")]
 pub fn latest_backup_migration_version() -> MigrationVersion {
     MIGRATIONS.get::<LatestBackupMigration>().unwrap().0.clone()
+}
+
+trait Convert {
+    type New: Serialize + DeserializeOwned;
+    type Old: DeserializeOwned;
+
+    fn check(value: &mut serde_json::Value) -> crate::wallet::Result<()> {
+        if serde_json::from_value::<Self::New>(value.clone()).is_err() {
+            *value = serde_json::to_value(Self::convert(serde_json::from_value::<Self::Old>(value.clone())?)?)?;
+        }
+        Ok(())
+    }
+
+    fn convert(old: Self::Old) -> crate::wallet::Result<Self::New>;
+}
+
+fn rename_keys(json: &mut serde_json::Value) {
+    match json {
+        serde_json::Value::Array(a) => a.iter_mut().for_each(rename_keys),
+        serde_json::Value::Object(o) => {
+            let mut replace = serde_json::Map::with_capacity(o.len());
+            o.retain(|k, v| {
+                rename_keys(v);
+                replace.insert(
+                    heck::ToLowerCamelCase::to_lower_camel_case(k.as_str()),
+                    std::mem::replace(v, serde_json::Value::Null),
+                );
+                true
+            });
+            *o = replace;
+        }
+        _ => (),
+    }
 }
